Move Binaryen specific code from eei.cpp to binaryen.cpp
